profile: editor: Add setting to preserve the working directory
The default value of the setting is 'nothing', which retains the
current behaviour of preserving the working directory only when the
command is a shell.

Therefore, unless explicitly changed, the 'preserve working directory'
check-button is kept synchronized with the 'use custom command'
check-button without altering the value of the underlying setting. To
make this work, an explicit signal handler is used to track changes to
the 'preserve working directory' check-button, as opposed to binding
it directly to the setting. This ensures that changes to the button
not initiated by user action don't modify the setting.

https://gitlab.gnome.org/GNOME/gnome-terminal/issues/126