window: Limit the title's length to avoid integer overflows in Pango
GtkLabels with really long texts of the order of 2^21 characters can
cause integer overflows in Pango when measuring the width of the
PangoLayout for that text. Tracking down the root cause of the overflow
in the font stack and fixing it is difficult, so it's prudent to work
around it by limiting the length of the window title to a sane, yet
large enough, number.

Pango does integer measurements in Pango units where 1 pixel is 2^10
Pango units, and font glyphs are roughly 16, or 2^4, pixels wide.
Therefore, roughly 2^(31-10-4), or 131072, characters is the threshold
after which Pango might start to overflow.

A limit of 1024 characters was chosen because that's the limit both
VTE and xterm use for titles set through the OSC 0 escape sequence.

https://gitlab.gnome.org/GNOME/gnome-terminal/issues/207